+++ JotunnLib/Configs/PieceConfig.cs
@@ -56,7 +56,7 @@ namespace Jotunn.Configs
         ///     The name of the crafting station prefab to which this piece will be an upgrade to.
         /// </summary>
         public string ExtendStation { get; set; } = string.Empty;
-        
+
         /// <summary>
         ///     Icon which is displayed in the crafting GUI.
         /// </summary>
@@ -82,19 +82,19 @@ namespace Jotunn.Configs
 
             piece.m_enabled = Enabled;
             piece.m_allowedInDungeons = AllowedInDungeons;
-            
+
             // Set name if given
             if (!string.IsNullOrEmpty(Name))
             {
                 piece.m_name = Name;
             }
-            
+
             // Set description if given
             if (!string.IsNullOrEmpty(Description))
             {
                 piece.m_description = Description;
             }
-            
+
             // Set icon if overriden
             if (Icon != null)
             {
@@ -132,14 +132,17 @@ namespace Jotunn.Configs
         /// <returns>The Valheim <see cref="Piece.Requirement"/> array</returns>
         public Piece.Requirement[] GetRequirements()
         {
-            Piece.Requirement[] reqs = new Piece.Requirement[Requirements.Length];
+            List<Piece.Requirement> reqs = new List<Piece.Requirement>();
 
-            for (int i = 0; i < reqs.Length; i++)
+            foreach (RequirementConfig requirement in Requirements)
             {
-                reqs[i] = Requirements[i].GetRequirement();
+                if (requirement != null && requirement.IsValid())
+                {
+                    reqs.Add(requirement.GetRequirement());
+                }
             }
 
-            return reqs;
+            return reqs.ToArray();
         }
 
         /// <summary>
@@ -163,12 +166,16 @@ namespace Jotunn.Configs
         }
 
         /// <summary>
-        ///     Appends a new <see cref="RequirementConfig"/> to the array of existing ones.
+        ///     Appends a new <see cref="RequirementConfig"/> to the array of existing ones.<br />
+        ///     If the requirement is null or is not valid (has not item name or amount set) nothing will be added.
         /// </summary>
         /// <param name="requirementConfig"></param>
         public void AddRequirement(RequirementConfig requirementConfig)
         {
-            Requirements = Requirements.AddToArray(requirementConfig);
+            if (requirementConfig != null && requirementConfig.IsValid())
+            {
+                Requirements = Requirements.AddToArray(requirementConfig);
+            }
         }
     }
 }