+++ JotunnLib/Configs/RequirementConfig.cs
@@ -61,5 +61,14 @@ namespace Jotunn.Configs
                 m_recover = Recover
             };
         }
+
+        /// <summary>
+        ///     Checks if the requirement has any item and amount set.
+        /// </summary>
+        /// <returns></returns>
+        internal bool IsValid()
+        {
+            return !string.IsNullOrEmpty(Item) && (Amount > 0 || AmountPerLevel > 0);
+        }
     }
 }